+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaClusterSpec.java
@@ -44,12 +44,14 @@ public class KafkaClusterSpec implements HasConfigurableMetrics, UnknownProperty
     public static final String FORBIDDEN_PREFIXES = "listeners, advertised., broker., listener., host.name, port, "
             + "inter.broker.listener.name, sasl., ssl., security., password., log.dir, "
             + "zookeeper.connect, zookeeper.set.acl, zookeeper.ssl, zookeeper.clientCnxnSocket, authorizer., super.user, "
-            + "cruise.control.metrics.topic, cruise.control.metrics.reporter.bootstrap.servers";
+            + "cruise.control.metrics.topic, cruise.control.metrics.reporter.bootstrap.servers,"
+            + "node.id, process.roles, controller."; // KRaft options
 
     public static final String FORBIDDEN_PREFIX_EXCEPTIONS = "zookeeper.connection.timeout.ms, ssl.cipher.suites, ssl.protocol, ssl.enabled.protocols,"
             + "cruise.control.metrics.topic.num.partitions, cruise.control.metrics.topic.replication.factor, cruise.control.metrics.topic.retention.ms,"
             + "cruise.control.metrics.topic.auto.create.retries, cruise.control.metrics.topic.auto.create.timeout.ms,"
-            + "cruise.control.metrics.topic.min.insync.replicas";
+            + "cruise.control.metrics.topic.min.insync.replicas,"
+            + "controller.quorum.election.backoff.max.ms, controller.quorum.election.timeout.ms, controller.quorum.fetch.timeout.ms"; // KRaft options
 
     protected Storage storage;
     private String version;